Finished Pickers and rearrange component file structure.
