--- assets/css/dashboard/new-pa-message-page.scss
@@ -1,280 +1,0 @@
-.new-pa-message-page {
-  padding: 24px 40px;
-  &__header {
-    font-size: 40px;
-    font-weight: 500;
-    line-height: 48px;
-    text-align: left;
-  }
-
-  &__associate-alert-subtext {
-    color: $text-secondary;
-  }
-
-  &__when,
-  &__where,
-  &__message {
-    margin-top: 24px;
-    padding: 18px 24px;
-    background-color: $cool-gray-20;
-    color: white;
-
-    .title {
-      margin-bottom: 16px;
-      font-weight: 500;
-      font-size: 24px;
-    }
-
-    .label {
-      margin-top: 16px;
-      margin-bottom: 8px;
-      color: $text-secondary;
-    }
-
-    .start-datetime,
-    .end-datetime {
-      width: 100%;
-      height: 38px;
-
-      form {
-        width: 250px;
-      }
-    }
-
-    .dropdown-toggle,
-    .form-control {
-      background-color: $dark-bg;
-      height: 38px;
-      color: $text-secondary;
-      font-weight: 400;
-      border: 1px solid $border-primary;
-    }
-
-    .dropdown-menu {
-      background-color: $cool-gray-30;
-    }
-
-    .dropdown-item {
-      color: $text-primary;
-      &:hover {
-        background-color: $cool-gray-20;
-      }
-
-      &.active {
-        background-color: $button-primary;
-        color: $button-secondary;
-      }
-    }
-
-    .interval {
-      width: 80px;
-      height: 38px;
-      .form-control {
-        padding: 7px 13px;
-        text-align: end;
-
-        border: 1px solid $border-primary;
-      }
-    }
-
-    input::-webkit-outer-spin-button,
-    input::-webkit-inner-spin-button {
-      /* display: none; <- Crashes Chrome on hover */
-      -webkit-appearance: none;
-      margin: 0;
-    }
-
-    input[type="number"] {
-      -moz-appearance: textfield; /* Firefox */
-      appearance: textfield;
-    }
-
-    .add-stations-zones-button {
-      background-color: $text-link-primary;
-      color: $text-invert;
-      border: 1px dashed $text-invert;
-      width: fit-content;
-      vertical-align: middle;
-    }
-
-    .copy-text-button {
-      background-color: $bg-border-disabled;
-      width: 38px;
-      height: 38px;
-
-      .bi-arrow-right-short {
-        fill: $text-disabled;
-      }
-    }
-
-    .visual-text-input,
-    .phonetic-text-input {
-      height: 250px;
-      resize: none;
-    }
-
-    .review-audio-card {
-      height: 250px;
-      background-color: $cool-gray-15;
-
-      .review-audio-button {
-        color: $button-primary;
-        margin-top: auto;
-        margin-bottom: auto;
-      }
-    }
-  }
-
-  &__form-buttons {
-    padding: 18px 24px;
-
-    .btn {
-      height: 38px;
-    }
-
-    .cancel-button {
-      background-color: transparent;
-      color: $text-link-primary;
-      border: none;
-    }
-
-    .submit-button {
-      background-color: $button-primary;
-      color: $button-secondary;
-    }
-  }
-}
-
-.calendar-tooltip {
-  margin-left: 12px;
-
-  .popover-arrow {
-    display: none;
-  }
-  .react-calendar {
-    padding: 8px 12px;
-    max-width: 250px;
-    background-color: $dark-bg;
-    border-radius: 4px;
-    border: 1px solid $text-disabled;
-    font-family: Inter;
-    line-height: 21px;
-    font-weight: 400;
-    font-size: 14px;
-
-    &__navigation {
-      height: 40px;
-      margin-bottom: 0;
-      button {
-        height: 100%;
-        &:enabled:hover,
-        &:enabled:focus,
-        &:disabled {
-          background-color: transparent;
-        }
-      }
-      &__label {
-        padding-left: 0;
-        padding-right: 0;
-      }
-      &__arrow {
-        padding: 0;
-        min-width: 22px;
-      }
-    }
-
-    &__viewContainer {
-      .react-calendar__month-view__weekdays {
-        color: $text-primary;
-        text-transform: unset;
-        flex: none;
-
-        &__weekday {
-          display: flex;
-          justify-content: center;
-          height: 30px;
-          width: 36px;
-          padding: 0;
-        }
-        abbr {
-          cursor: unset;
-          text-decoration: unset;
-          margin: auto;
-        }
-      }
-
-      .react-calendar__month-view__days {
-        .react-calendar__tile {
-          height: 30px;
-          width: 36px;
-          padding: 1.5px 4px;
-          border-radius: 4px;
-          &--now {
-            background-color: transparent;
-            border: 1px solid $text-disabled;
-            color: $text-primary;
-
-            &:enabled:hover {
-              background-color: $main-blue;
-              border-color: $button-primary;
-            }
-          }
-
-          &--active {
-            background-color: $button-primary;
-            color: $cool-gray-15;
-            border: none;
-          }
-
-          &:enabled:hover {
-            background-color: $cool-gray-20;
-          }
-          &:disabled {
-            background-color: transparent;
-            color: $text-disabled;
-          }
-        }
-
-        &__day {
-          &--weekend {
-            color: $text-primary;
-          }
-          &--neighboringMonth {
-            color: $text-disabled;
-          }
-        }
-      }
-    }
-  }
-}
-
-.time-tooltip {
-  margin-left: 12px;
-
-  .time-picker-card {
-    width: 216px;
-    overflow: hidden;
-  }
-
-  .col {
-    padding: 0;
-    overflow-y: scroll;
-  }
-
-  .hour-col,
-  .minute-col {
-    .btn {
-      width: 100%;
-      border-radius: 0;
-    }
-  }
-
-  .popover-arrow {
-    display: none;
-  }
-
-  .row {
-    max-height: 417px;
-    flex-wrap: nowrap;
-  }
-}
