+++ assets/css/dashboard/new-pa-message-page/message-card.scss
@@ -1,0 +1,47 @@
+.message-card {
+  margin-top: 24px;
+  padding: 18px 24px;
+  background-color: $cool-gray-20;
+  color: white;
+
+  .title {
+    margin-bottom: 16px;
+    font-weight: 500;
+    font-size: 24px;
+  }
+
+  .form-control {
+    background-color: $dark-bg;
+    height: 38px;
+    color: $text-secondary;
+    font-weight: 400;
+    border: 1px solid $border-primary;
+  }
+
+  .copy-text-button {
+    background-color: $bg-border-disabled;
+    width: 38px;
+    height: 38px;
+
+    .bi-arrow-right-short {
+      fill: $text-disabled;
+    }
+  }
+
+  .visual-text-input,
+  .phonetic-text-input {
+    height: 250px;
+    resize: none;
+  }
+
+  .review-audio-card {
+    height: 250px;
+    background-color: $cool-gray-15;
+
+    .review-audio-button {
+      color: $button-primary;
+      margin-top: auto;
+      margin-bottom: auto;
+    }
+  }
+}
