+++ assets/css/dashboard/new-pa-message-page/when-card.scss
@@ -1,0 +1,76 @@
+.when-card {
+  margin-top: 24px;
+  padding: 18px 24px;
+  background-color: $cool-gray-20;
+  color: white;
+
+  .title {
+    margin-bottom: 16px;
+    font-weight: 500;
+    font-size: 24px;
+  }
+
+  .label {
+    margin-top: 16px;
+    margin-bottom: 8px;
+    color: $text-secondary;
+  }
+
+  .start-datetime,
+  .end-datetime {
+    width: 100%;
+    height: 38px;
+
+    form {
+      width: 250px;
+    }
+  }
+
+  .dropdown-toggle,
+  .form-control {
+    background-color: $dark-bg;
+    height: 38px;
+    color: $text-secondary;
+    font-weight: 400;
+    border: 1px solid $border-primary;
+  }
+
+  .dropdown-menu {
+    background-color: $cool-gray-30;
+  }
+
+  .dropdown-item {
+    color: $text-primary;
+    &:hover {
+      background-color: $cool-gray-20;
+    }
+
+    &.active {
+      background-color: $button-primary;
+      color: $button-secondary;
+    }
+  }
+
+  .interval {
+    width: 80px;
+    height: 38px;
+    .form-control {
+      padding: 7px 13px;
+      text-align: end;
+
+      border: 1px solid $border-primary;
+    }
+  }
+
+  input::-webkit-outer-spin-button,
+  input::-webkit-inner-spin-button {
+    /* display: none; <- Crashes Chrome on hover */
+    -webkit-appearance: none;
+    margin: 0;
+  }
+
+  input[type="number"] {
+    -moz-appearance: textfield; /* Firefox */
+    appearance: textfield;
+  }
+}
