+++ assets/css/dashboard/time-picker.scss
@@ -1,0 +1,103 @@
+.time-popover {
+  margin-left: 12px;
+
+  .time-picker-card {
+    width: 216px;
+    overflow: hidden;
+    background-color: $dark-bg;
+    border: 1px solid $text-disabled;
+    font-family: Inter;
+
+    .time-picker-card-columns {
+      display: flex;
+      flex-direction: row;
+      max-height: 400px;
+
+      .hour-col,
+      .minute-col {
+        border-right: 1px solid $border-primary;
+      }
+
+      .hour-col,
+      .minute-col,
+      .ampm-col {
+        padding: 0;
+        overflow-y: auto;
+        display: flex;
+        flex-direction: column;
+        flex-shrink: 0;
+        flex-grow: 1;
+
+        scrollbar-width: thin;
+      }
+
+      button {
+        font-size: 14px;
+        font-weight: 400;
+        margin: 0;
+        padding: 0;
+        width: auto;
+        background: none;
+        border: none;
+        line-height: 0;
+        height: 28px;
+        padding: 3.5px 24px 3.5px 14px;
+        flex-shrink: 0;
+
+        &:hover {
+          background-color: $cool-gray-30;
+        }
+
+        &.selected {
+          background-color: $button-primary;
+          color: $text-invert;
+        }
+      }
+
+      .ampm-col {
+        button {
+          padding: 3.5px 67pc 3.5px 14px;
+        }
+      }
+    }
+
+    .bottom-buttons {
+      height: 76px;
+      padding: 9px 12px;
+      border-top: 1px solid $border-primary;
+      display: flex;
+      align-items: center;
+
+      button {
+        border: none;
+        height: fit-content;
+      }
+
+      .start-of-service-button,
+      .end-of-service-button {
+        font-size: 16px;
+        font-weight: 400;
+        padding: 0;
+
+        background-color: transparent;
+        color: $button-primary;
+      }
+
+      .start-of-service-button {
+        margin-bottom: 10px;
+      }
+
+      .ok-button {
+        font-size: 14px;
+        background-color: $button-primary;
+        color: $dark-bg;
+        padding: 4px 8px;
+        border-radius: 4px;
+      }
+    }
+  }
+
+  .popover-arrow {
+    display: none;
+  }
+}
