+++ assets/css/screenplay.scss
@@ -23,7 +23,10 @@ $form-check-input-border: 1px solid #8b9198;
 @import "dashboard/bottom-action-bar.scss";
 @import "dashboard/configure-screens-workflow-page.scss";
 @import "dashboard/pa-messages-page.scss";
-@import "dashboard/new-pa-message-page.scss";
+@import "dashboard/new-pa-message-page/new-pa-message-page.scss";
+@import "dashboard/new-pa-message-page/when-card.scss";
+@import "dashboard/new-pa-message-page/where-card.scss";
+@import "dashboard/new-pa-message-page/message-card.scss";
 @import "place-row.scss";
 @import "places-action-bar.scss";
 @import "filter-dropdown.scss";
@@ -41,6 +44,8 @@ $form-check-input-border: 1px solid #8b9198;
 @import "workflow.scss";
 @import "sort-label.scss";
 @import "search-bar.scss";
+@import "dashboard/date-picker.scss";
+@import "dashboard/time-picker.scss";
 
 html {
   font-size: 16px;
